api: add region label api (#3984)
* labeler: add region label support


* fix comment


* error doc


* api: add region label api


* fix error code doc


* update swagger document

+++ server/schedule/labeler/labeler.go
@@ -192,6 +192,9 @@ func (l *RegionLabeler) SetLabelRule(rule *LabelRule) error {
 func (l *RegionLabeler) DeleteLabelRule(id string) error {
 	l.Lock()
 	defer l.Unlock()
+	if _, ok := l.labelRules[id]; !ok {
+		return errs.ErrRegionRuleNotFound.FastGenByArgs(id)
+	}
 	if err := l.storage.DeleteRegionRule(id); err != nil {
 		return err
 	}
